Get rid of an error

+++ src/Test/conftest.py
@@ -109,13 +109,14 @@ from Debug import Debug
 def cleanup():
     Db.dbCloseAll()
     for dir_path in [config.data_dir, config.data_dir + "-temp"]:
-        for file_name in os.listdir(dir_path):
-            ext = file_name.rsplit(".", 1)[-1]
-            if ext not in ["csr", "pem", "srl", "db", "json", "tmp"]:
-                continue
-            file_path = dir_path + "/" + file_name
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
+        if os.path.isdir(dir_path):
+            for file_name in os.listdir(dir_path):
+                ext = file_name.rsplit(".", 1)[-1]
+                if ext not in ["csr", "pem", "srl", "db", "json", "tmp"]:
+                    continue
+                file_path = dir_path + "/" + file_name
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
 
 atexit.register(cleanup)
